Consolidate authentication bits into Session
